Various minor fixes.
* include/mailutils/diag.h (mu_full_program_name): New extern.
* libmailutils/diag/diag.c (mu_full_program_name): New global.
(mu_set_program_name): Set mu_full_program_name.
(mu_diag_at_locus): Fix format specifier.
* libmailutils/diag/bt.c: Use mu_full_program_name instead of
program_invocation_name.
* libmailutils/diag/wd.c (mu_wd): Likewise.

* comsat/comsat.c: Fix format specifiers.
* lib/mailcap.c: Likewise.
* libmu_sieve/util.c: Cast formatted arguments to match format specs.
* mh/mh_getopt.c: Likewise.

* libmailutils/cidr/fromstr.c: Add missing includes.
* libproto/mailer/smtp.c: Likewise.
* libproto/pop/mbox.c: Likewise.
* testsuite/smtpsend.c: Likewise.
* mu/getarg.c: Likewise.
* mu/imap.c: Likewise.
* mu/pop.c (com_connect): Protect TLS code with #ifdef.
* mu/Makefile.am (EXTRA_DIST): Add mu-setup.awk.

* mail/summary.c (mail_summary): Fix data types of the counters.

+++ comsat/comsat.c
@@ -379,9 +379,10 @@ comsat_connection (int fd, struct sockaddr *sa, int salen,
     {
       char *p = mu_sockaddr_to_astr (sa, salen);
       mu_diag_output (MU_DIAG_INFO,
-		      ngettext ("received %d byte from %s",
-				"received %d bytes from %s", rdlen),
-		      rdlen, p);
+		      ngettext ("received %lu byte from %s",
+				"received %lu bytes from %s",
+				(unsigned long) rdlen),
+		      (unsigned long) rdlen, p);
       mu_diag_output (MU_DIAG_INFO, "string: %s", buffer);
       free (p);
     }
